Move utils from Frontend.Handler to Frontend.Util.
